Improve video scaling/aspect ratios
A few changes:
- Swap aspect's subprocess.getoutput with subprocess.run
- Calculate aspect ratio instead of relying on mplayer's aspect ratio
ID. Some videos did not work with the aspect ratio ID (ID showed an
AR of 0.0000...), so this is an improvement to ensure all videos work.
- Use the expand option in mencoders video filters to ensure that
the video will not be smaller than 256x192. Width is important as
Moonshell will not play the video otherwise, height doesn't seem as
important but it'd be best to show the whole video on screen.
Expand is different to scale as it will not alter the videos aspect
ratio/size.