Remove mpeg_stat requirement
This was the main hurdle for dpgconv - a requirement that was very hard
to find and replace. I did not want to introduce a new requirement just
so that this wouldn't happen again in future for it's replacement.

As we have set the keyframe interval in lavc flags to be 10 frames, we
know that our sequence is going to comprise of 10 pictures. All we need
to do is check for the sequence start code, then the start codes for
the rest of the 10 pictures. If either returns with nothing, we have
reached the end of the file and know not to continue.

The previous behaviour with DPG2 GOP header is kept in place, instead
of getting the output first then figuring out the GOP header, it is
created while the output is being made... if that makes sense. All in
all, with my limited testing, seeking is still as fast as before so I
think this is good enough.